+++ android/app/src/main/java/net/melisma/selko/SelkoApplication.kt
@@ -1,0 +1,20 @@
+package net.melisma.selko
+
+import android.app.Application
+import net.melisma.selko.di.appModule
+import org.koin.android.ext.koin.androidContext
+import org.koin.android.ext.koin.androidLogger
+import org.koin.core.context.startKoin
+import org.koin.core.logger.Level
+
+class SelkoApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        startKoin {
+            androidLogger(Level.ERROR)
+            androidContext(this@SelkoApplication)
+            modules(appModule)
+        }
+    }
+}