+++ android/app/src/main/java/net/melisma/selko/ui/navigation/NavRoutes.kt
@@ -1,0 +1,9 @@
+package net.melisma.selko.ui.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+object Auth
+
+@Serializable
+object Home